server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-dotenv==1.0.1
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0
google-generativeai==0.8.3
httpx==0.27.2
pydantic==2.9.0
pydantic-settings==2.5.2
aiofiles==24.1.0
uuid7==0.1.0
